Changed fuel breakout to all fuels; monthly MWh
The breakout of tech opportunity by fuel type has been changed to represent the monthly sum of all fuels displaced by the solar technology package. The hourly fuel mix is used with the hourly technical opportunity and hourly total county load to calculate the
monthly MWh of fuels displaced by county. The option to select a subset of fuels remains. The H5 file script was also changed to reflect the fuel breakout is now in MWh on a monthly basis. A month index array was added to the H5 file. FYI @carriescho, @wxi1995