Fixed additional bugs for breaking out all fuels
When running the PTC tech opp, found an error related to the recalculation of county-level
data. Realized that `mecs_fuel_intensity.csv` had not been updated with the revised
county-level data. Also made a small change to the way that `tech_opp_demand.py` accounts for
fuel types not used in a county.