IndexError in tech_opp_demand.py
Working through a fix to an IndexError exception that occurs when
constructing a
pivot table from a an empty dataframe. It appears that this might be
caused by counties that have only GHGRP reporters and no smaller
facilities. Changed the merging and concatentating operations when
constructing the fuel_dfs dataframes from the mecs and GHGRP fuel
intensity dataframes. Also added a temporary exception handling of the
IndexError to identify which counties are problematic. Wasn't able to
identify these otherwise. An example is 38007, which contains only a
single 324199 GHGRP reporter.