Updating land area estimates
An earlier mistake was made when estimating the available land area by county. The exclusion rule for
slope has been changed from 10% to 5%. The input file to format_rev_output.py was added to the
calculation_data directory and the format_rev_output.py code changed to reference this new input file.